add an events file and move all events-related code to the file
